requirements.txt
requests
beautifulsoup4
python-dotenv
pyjwt[crypto]
